[VGNWB-366] re-organized relation helper methods, fixed text view relation creating/updating
