+++ annotations/model_utils.py
@@ -1,0 +1,8 @@
+import re
+
+def help_text(text):
+    """
+    Remove excess whitespace from a string. Intended for use in model and form
+    fields when writing long help_texts.
+    """
+    return re.sub('(\s+)', ' ', text)
